+++ client-portals/components/maintainer/dashboard/index.tsx
@@ -1,0 +1,2 @@
+export { default as Card } from "../../shared/card";
+export { default as Dashboard } from "./dashboard";
