--- client-portals/components/portals/maintainer/dashboard/index.tsx
@@ -1,2 +1,0 @@
-export { default as Card } from "../../../shared/card";
-export { default as Dashboard } from "./dashboard";
