+++ client-portals/components/shared/index.tsx
@@ -1,1 +1,2 @@
 export { default as Input } from "./input";
+export { default as Layout } from "./layout";
